Speedrun search is now working Additional information is now displayed to the right of the speedrun list

+++ src/main/java/com/bizcub/randomItemSpeedrun/gui/Speedrun.java
@@ -6,6 +6,6 @@ public record Speedrun(
         ItemStack itemStack,
         boolean isSuccess,
         int time,
-        int date
+        long date
 ) {
 }
